audio_signal - ENH - rate is recasted as float and data is recasted as float32
Former-commit-id: 49c52ad917f86945f0d88b258429ade9cafc7bbd
Former-commit-id: 42ef6c00acc877b178663cc2322f5b225e0744c5

+++ sound_classification/audio_signal.py
@@ -8,14 +8,14 @@ class AudioSignal:
     """ Audio signal
 
         Args:
-            rate: int
+            rate: float
                 Sampling rate in Hz
             data: 1d numpy array
                 Audio data 
     """
     def __init__(self, rate, data, tag=""):
-        self.rate = rate
-        self.data = data
+        self.rate = float(rate)
+        self.data = data.astype(dtype=np.float32)
         self.tag = tag
 
     @classmethod
